Python Programs/Let me in/requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cmake==3.18.2.post1
colorama==0.4.3
cryptography==3.1.1
cx-Freeze==6.2
decorator==4.4.2
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.10
idna-ssl==1.1.0
imutils==0.5.3
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
numpy==1.19.2
opencv-contrib-python==4.4.0.44
pandas==1.1.2
Pillow==7.2.0
pycparser==2.20
pylint==2.6.0
pyserial==3.4
python-dateutil==2.8.1
python-telegram-bot==12.8
pytz==2020.1
six==1.15.0
telepot==12.7
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wrapt==1.12.1
yarl==1.6.0
